Initial commit, with some useful code
The current code in scripts/gen-most.sh and implemented by 'lein run
json2edn' helps automate the steps to run 'lein clojuredocs' on
Clojure and most of the current contrib libraries.

'lein run cdocs-summary' summarizes the namespaces and number of
public and private vars defined in each of the projects that the above
step has been performed on.